1.1.0 (#5)
* perf: optimize metamodel

* refactor: adaptation jdk 8 syntax

* refactor: adaptation jdk 8 syntax

* feat: Support for mapping private fields

* feat: Support `orderBy` method streaming build

* refactor: refactor SQL printing, cancel printing empty parameters

* feat: modify modifier to support extensions

* fix: fix build Metamodel may contains transient fields

* refactor: simplify `Metamodel` implementation code

* feat: add ExpressionBuilder api

* feat: add ExpressionBuilder api and implements

* feat: add ExpressionBuilder api and implements

* feat: separate expression operations

* feat: reorganize implementation classes and standardize naming

* feat: Supports superclass inheritance properties with 'MappedSuperclass' annotation

* feat: Supports superclass inheritance properties with 'MappedSuperclass' annotation

* feat: support select distinct query

* style: delete redundant comments

* feat: support join projection attribute query

* feat: add the type check of the field annotated with `EntityAttribute`

* style: rename some variables

* feat: change the return value of count function to long

* test: add test cases

* test: add test cases

* feat: Support for nested projections

* feat: Support for nested projections

* perf: improves performance

* perf: improves performance

* perf: improves performance

* test: add test cases

* refactor: separate the reflection code

* style: removing extra and formatting code

* build: update version to 1.1.0-SNAPSHOT

* refactor: refactoring `Metamodel` implementation

* refactor: `ColumnImpl` override equals and hashCode

* 增加查询方法，增加Updater接口，优化测试代码

* 修复: in参数列表长度为1时查询结果为空; having参数不生效

* 增加测试用例

* 增加测试用例

* 增加whereIf查询方法

* 增加andIf,ofIf查询方法,重命名EntityRoot为Root

* 删除不必要的接口Query.AndBuilder,Query.OrBuilder

* 删除工具类Util增加类PathReference

* Object[]查询结果改为Tuple

* 增加测试用例

* 增加测试用例

* 增加测试用例

* 模块化

* 增加maven.compiler.release属性

* 修改测试用例

* 添加lombok配置

* 删除io.github.genie.sql.builder.meta.Type#name方法

* 删除module-info多余的exports

* 更新版本号

---------

+++ sql-genie-api/src/main/java/io/github/genie/sql/api/Sliceable.java
@@ -8,6 +8,6 @@ public interface Sliceable<T, U> {
 
     int limit();
 
-    U collect(List<T> list, int total);
+    U collect(List<T> list, long total);
 
 }
